Add files via upload
Analysing the following game:
A hat contains 2024 pieces of paper numbered 1 through 2024. I draw two pieces of paper at random from the hat. The smaller of the two numbers drawn is subtracted from the larger. That difference is written on a new piece of paper which is placed in the hat, and the two original pieces are discarded. I repeat this process until one piece of paper remains. 

What can you tell about the final piece of paper?