Add files via upload
Used a recursion solver to help solve the following problem:

You have a fair coin which has the number 1 written on one side and the number 2 on the other. You throw the coin, and if it comes up N, you would then take N steps to the right. 
For example, if you throw the coin and it comes up 2, you take 2 steps to the right to land on step number 3.
You now repeat the exercise, throwing the coin again and walking the number of steps that comes up on the coin. If you throw the coin 24 times, you are certain to have landed on, or past, step number 25. 

What is the probability that at some point in this exercise you will land on step number 25?